packages/aserto-idp/pyproject.toml
[tool.poetry]
name = "aserto-idp"
version = "0.3.0"
description = "Common identity providers for use with Aserto client libraries"
readme = "README.md"
authors = ["Aserto, Inc. <pypi@aserto.com>"]
maintainers = ["authereal <authereal@aserto.com>"]
homepage = "https://github.com/aserto-dev/aserto-python/tree/HEAD/packages/aserto-idp"
repository = "https://github.com/aserto-dev/aserto-python/tree/HEAD/packages/aserto-idp"
documentation = "https://github.com/aserto-dev/aserto-python/tree/HEAD/packages/aserto-idp"
license = "Apache-2.0"
classifiers = [
  "Intended Audience :: Developers",
  "License :: OSI Approved :: Apache Software License",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Topic :: Software Development :: Libraries",
  "Typing :: Typed",
]
packages = [
    {include = "aserto_idp", from = "src"}
]

[tool.poetry.dependencies]
aiohttp = "^3.8.1"
python = "^3.7"
python-jose = {version = "^3.3.0", extras = ["cryptography"]}
typing-extensions = "^3.10.0"

[tool.poetry.dev-dependencies]
black = "21.7b0"
isort= "^5.9.3"
mypy = "^0.910"
pytest-asyncio = "^0.15.1"
pytest-cov = "^3.0.0"

[tool.black]
line-length = 100
target-version = ["py37"]

[tool.isort]
profile = "black"

[tool.mypy]
show_error_codes = true
check_untyped_defs = true
disallow_any_decorated = true
disallow_any_explicit = true
disallow_any_expr = true
disallow_any_generics = true
disallow_any_unimported = true
disallow_incomplete_defs = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_decorators = true
disallow_untyped_defs = true
implicit_reexport = false
no_implicit_optional = true
strict_equality = true
strict_optional = true
warn_no_return = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = ["jose"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
# jose generates too many `Any` types
module = ["aserto_idp.auth0", "aserto_idp.oidc.*"]
disallow_any_expr = false

[[tool.mypy.overrides]]
# lru_cache swallows typing info
module = ["aserto_idp.oidc.discovery"]
warn_return_any = false
disallow_untyped_decorators = false
disallow_any_decorated = false

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

